chore: Configure DWI with tlsCertificateConfigmapRef when certificates imported to che-operator
