--- pkg/deploy/devworkspace/dev_workspace_config_test.go
@@ -1,3780 +1,0 @@
-//
-// Copyright (c) 2019-2025 Red Hat, Inc.
-// This program and the accompanying materials are made
-// available under the terms of the Eclipse Public License 2.0
-// which is available at https://www.eclipse.org/legal/epl-2.0/
-//
-// SPDX-License-Identifier: EPL-2.0
-//
-// Contributors:
-//   Red Hat, Inc. - initial API and implementation
-//
-
-package devworkspace
-
-import (
-	"context"
-	"fmt"
-	"regexp"
-	"sort"
-	"testing"
-
-	"k8s.io/utils/ptr"
-	"sigs.k8s.io/controller-runtime/pkg/client"
-
-	controllerv1alpha1 "github.com/devfile/devworkspace-operator/apis/controller/v1alpha1"
-	chev2 "github.com/eclipse-che/che-operator/api/v2"
-	"github.com/eclipse-che/che-operator/pkg/common/constants"
-	"github.com/eclipse-che/che-operator/pkg/common/test"
-	"github.com/eclipse-che/che-operator/pkg/deploy/tls"
-	"github.com/google/go-cmp/cmp"
-	"github.com/google/go-cmp/cmp/cmpopts"
-	"github.com/stretchr/testify/assert"
-	corev1 "k8s.io/api/core/v1"
-	"k8s.io/apimachinery/pkg/api/resource"
-	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
-	"k8s.io/apimachinery/pkg/types"
-)
-
-func TestReconcileDevWorkspaceConfigStorage(t *testing.T) {
-	type testCase struct {
-		name                   string
-		cheCluster             *chev2.CheCluster
-		existedObjects         []client.Object
-		expectedOperatorConfig *controllerv1alpha1.OperatorConfiguration
-	}
-
-	type errorTestCase struct {
-		name                 string
-		cheCluster           *chev2.CheCluster
-		existedObjects       []client.Object
-		expectedErrorMessage string
-	}
-
-	var quantity15Gi = resource.MustParse("15Gi")
-	var quantity10Gi = resource.MustParse("10Gi")
-
-	var expectedErrorTestCases = []errorTestCase{
-		{
-			name: "Invalid claim size string",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						Storage: chev2.WorkspaceStorage{
-							PvcStrategy: constants.PerUserPVCStorageStrategy,
-							PerUserStrategyPvcConfig: &chev2.PVC{
-								StorageClass: "test-storage",
-								ClaimSize:    "invalid-ClaimSize",
-							},
-						},
-					},
-				},
-			},
-			expectedErrorMessage: "quantities must match the regular expression",
-		},
-	}
-
-	var testCases = []testCase{
-		{
-			name: "Create DevWorkspaceOperatorConfig",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{Workspace: &controllerv1alpha1.WorkspaceConfig{DeploymentStrategy: "Recreate"}},
-		},
-		{
-			name: "Create DevWorkspaceOperatorConfig with ephemeral storage strategy",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						Storage: chev2.WorkspaceStorage{
-							PvcStrategy: constants.EphemeralPVCStorageStrategy,
-							PerUserStrategyPvcConfig: &chev2.PVC{
-								StorageClass: "test-storage",
-								ClaimSize:    "10Gi",
-							},
-							PerWorkspaceStrategyPvcConfig: &chev2.PVC{
-								StorageClass: "test-storage",
-								ClaimSize:    "10Gi",
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{Workspace: &controllerv1alpha1.WorkspaceConfig{DeploymentStrategy: "Recreate"}},
-		},
-		{
-			name: "Create DevWorkspaceOperatorConfig with StorageClassName only",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						Storage: chev2.WorkspaceStorage{
-							PvcStrategy: constants.PerUserPVCStorageStrategy,
-							PerUserStrategyPvcConfig: &chev2.PVC{
-								StorageClass: "test-storage",
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					StorageClassName:   ptr.To("test-storage"),
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Create DevWorkspaceOperatorConfig with non empty StorageAccessMode",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						Storage: chev2.WorkspaceStorage{
-							PvcStrategy: constants.PerUserPVCStorageStrategy,
-							PerUserStrategyPvcConfig: &chev2.PVC{
-								StorageAccessMode: []corev1.PersistentVolumeAccessMode{
-									corev1.ReadWriteMany,
-								},
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					StorageAccessMode: []corev1.PersistentVolumeAccessMode{
-						corev1.ReadWriteMany,
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Not setting PerUserStrategyPvcConfig should reset DevWorkspaceConfig to default StorageAccessMode",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						Storage: chev2.WorkspaceStorage{
-							PvcStrategy: constants.PerUserPVCStorageStrategy,
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Create DevWorkspaceOperatorConfig with nil StorageAccessMode",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						Storage: chev2.WorkspaceStorage{
-							PvcStrategy: constants.PerUserPVCStorageStrategy,
-							PerUserStrategyPvcConfig: &chev2.PVC{
-								StorageAccessMode: nil,
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					StorageAccessMode:  nil,
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Create DevWorkspaceOperatorConfig with non empty default container resources",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						DefaultContainerResources: &corev1.ResourceRequirements{
-							Limits: corev1.ResourceList{
-								corev1.ResourceMemory: resource.MustParse("128Mi"),
-							},
-							Requests: corev1.ResourceList{
-								corev1.ResourceMemory: resource.MustParse("64Mi"),
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					DefaultContainerResources: &corev1.ResourceRequirements{
-						Limits: corev1.ResourceList{
-							corev1.ResourceMemory: resource.MustParse("128Mi"),
-						},
-						Requests: corev1.ResourceList{
-							corev1.ResourceMemory: resource.MustParse("64Mi"),
-						},
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Create DevWorkspaceOperatorConfig with nil default container resources",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						DefaultContainerResources:         nil,
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					DefaultContainerResources: nil,
-					DeploymentStrategy:        "Recreate",
-				},
-			},
-		},
-		{
-			name: "Create DevWorkspaceOperatorConfig with non empty container resource caps",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						ContainerResourceCaps: &corev1.ResourceRequirements{
-							Limits: corev1.ResourceList{
-								corev1.ResourceMemory: resource.MustParse("2Gi"),
-								corev1.ResourceCPU:    resource.MustParse("2000m"),
-							},
-							Requests: corev1.ResourceList{
-								corev1.ResourceMemory: resource.MustParse("1Gi"),
-								corev1.ResourceCPU:    resource.MustParse("1000m"),
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ContainerResourceCaps: &corev1.ResourceRequirements{
-						Limits: corev1.ResourceList{
-							corev1.ResourceMemory: resource.MustParse("2Gi"),
-							corev1.ResourceCPU:    resource.MustParse("2000m"),
-						},
-						Requests: corev1.ResourceList{
-							corev1.ResourceMemory: resource.MustParse("1Gi"),
-							corev1.ResourceCPU:    resource.MustParse("1000m"),
-						},
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Create DevWorkspaceOperatorConfig with nil container resource caps",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						ContainerResourceCaps:             nil,
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ContainerResourceCaps: nil,
-					DeploymentStrategy:    "Recreate",
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig by adding container resource caps",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						ContainerResourceCaps: &corev1.ResourceRequirements{
-							Limits: corev1.ResourceList{
-								corev1.ResourceMemory: resource.MustParse("2Gi"),
-								corev1.ResourceCPU:    resource.MustParse("2000m"),
-							},
-							Requests: corev1.ResourceList{
-								corev1.ResourceMemory: resource.MustParse("1Gi"),
-								corev1.ResourceCPU:    resource.MustParse("1000m"),
-							},
-						},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							ContainerResourceCaps: nil,
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ContainerResourceCaps: &corev1.ResourceRequirements{
-						Limits: corev1.ResourceList{
-							corev1.ResourceMemory: resource.MustParse("2Gi"),
-							corev1.ResourceCPU:    resource.MustParse("2000m"),
-						},
-						Requests: corev1.ResourceList{
-							corev1.ResourceMemory: resource.MustParse("1Gi"),
-							corev1.ResourceCPU:    resource.MustParse("1000m"),
-						},
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig by changing container resource caps",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						ContainerResourceCaps: &corev1.ResourceRequirements{
-							Limits: corev1.ResourceList{
-								corev1.ResourceMemory: resource.MustParse("4Gi"),
-								corev1.ResourceCPU:    resource.MustParse("4000m"),
-							},
-							Requests: corev1.ResourceList{
-								corev1.ResourceMemory: resource.MustParse("2Gi"),
-								corev1.ResourceCPU:    resource.MustParse("2000m"),
-							},
-						},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							ContainerResourceCaps: &corev1.ResourceRequirements{
-								Limits: corev1.ResourceList{
-									corev1.ResourceMemory: resource.MustParse("2Gi"),
-									corev1.ResourceCPU:    resource.MustParse("2000m"),
-								},
-								Requests: corev1.ResourceList{
-									corev1.ResourceMemory: resource.MustParse("1Gi"),
-									corev1.ResourceCPU:    resource.MustParse("1000m"),
-								},
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ContainerResourceCaps: &corev1.ResourceRequirements{
-						Limits: corev1.ResourceList{
-							corev1.ResourceMemory: resource.MustParse("4Gi"),
-							corev1.ResourceCPU:    resource.MustParse("4000m"),
-						},
-						Requests: corev1.ResourceList{
-							corev1.ResourceMemory: resource.MustParse("2Gi"),
-							corev1.ResourceCPU:    resource.MustParse("2000m"),
-						},
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig by removing container resource caps",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						ContainerResourceCaps:             nil,
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							ContainerResourceCaps: &corev1.ResourceRequirements{
-								Limits: corev1.ResourceList{
-									corev1.ResourceMemory: resource.MustParse("2Gi"),
-									corev1.ResourceCPU:    resource.MustParse("2000m"),
-								},
-								Requests: corev1.ResourceList{
-									corev1.ResourceMemory: resource.MustParse("1Gi"),
-									corev1.ResourceCPU:    resource.MustParse("1000m"),
-								},
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ContainerResourceCaps: nil,
-					DeploymentStrategy:    "Recreate",
-				},
-			},
-		},
-		{
-			name: "Create DevWorkspaceOperatorConfig with per-user storage strategy",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						Storage: chev2.WorkspaceStorage{
-							PvcStrategy: constants.PerUserPVCStorageStrategy,
-							PerUserStrategyPvcConfig: &chev2.PVC{
-								StorageClass: "test-storage",
-								ClaimSize:    "15Gi",
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					StorageClassName: ptr.To("test-storage"),
-					DefaultStorageSize: &controllerv1alpha1.StorageSizes{
-						Common: &quantity15Gi,
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Create DevWorkspaceOperatorConfig with per-workspace storage strategy",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						Storage: chev2.WorkspaceStorage{
-							PvcStrategy: constants.PerWorkspacePVCStorageStrategy,
-							PerWorkspaceStrategyPvcConfig: &chev2.PVC{
-								StorageClass: "test-storage",
-								ClaimSize:    "15Gi",
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					StorageClassName: ptr.To("test-storage"),
-					DefaultStorageSize: &controllerv1alpha1.StorageSizes{
-						PerWorkspace: &quantity15Gi,
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Update DevWorkspaceOperatorConfig with per-workspace storage strategy",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						Storage: chev2.WorkspaceStorage{
-							PvcStrategy: constants.PerWorkspacePVCStorageStrategy,
-							PerWorkspaceStrategyPvcConfig: &chev2.PVC{
-								StorageClass: "test-storage",
-								ClaimSize:    "15Gi",
-							},
-						},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							StorageClassName: ptr.To("default-storage-class"),
-							DefaultStorageSize: &controllerv1alpha1.StorageSizes{
-								PerWorkspace: &quantity10Gi,
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					StorageClassName: ptr.To("test-storage"),
-					DefaultStorageSize: &controllerv1alpha1.StorageSizes{
-						PerWorkspace: &quantity15Gi,
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Update DevWorkspaceOperatorConfig with per-user storage strategy",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						Storage: chev2.WorkspaceStorage{
-							PvcStrategy: constants.PerUserPVCStorageStrategy,
-							PerUserStrategyPvcConfig: &chev2.PVC{
-								StorageClass: "test-storage",
-								ClaimSize:    "15Gi",
-							},
-						},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							StorageClassName: ptr.To("default-storage-class"),
-							DefaultStorageSize: &controllerv1alpha1.StorageSizes{
-								Common: &quantity10Gi,
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					StorageClassName: ptr.To("test-storage"),
-					DefaultStorageSize: &controllerv1alpha1.StorageSizes{
-						Common: &quantity15Gi,
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Update populated DevWorkspaceOperatorConfig with storage class name, storage strategy and storage size",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						Storage: chev2.WorkspaceStorage{
-							PvcStrategy: constants.PerUserPVCStorageStrategy,
-							PerUserStrategyPvcConfig: &chev2.PVC{
-								StorageClass: "test-storage",
-								ClaimSize:    "15Gi",
-							},
-						},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Routing: &controllerv1alpha1.RoutingConfig{
-							DefaultRoutingClass: "routing-class",
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Routing: &controllerv1alpha1.RoutingConfig{
-					DefaultRoutingClass: "routing-class",
-				},
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					StorageClassName: ptr.To("test-storage"),
-					DefaultStorageSize: &controllerv1alpha1.StorageSizes{
-						Common: &quantity15Gi,
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).WithObjects(testCase.existedObjects...).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(context.TODO(), types.NamespacedName{Name: devWorkspaceConfigName, Namespace: testCase.cheCluster.Namespace}, dwoc)
-			assert.NoError(t, err)
-
-			diff := cmp.Diff(testCase.expectedOperatorConfig, dwoc.Config, cmp.Options{
-				cmpopts.IgnoreFields(controllerv1alpha1.WorkspaceConfig{}, "ServiceAccount"),
-				cmpopts.IgnoreFields(controllerv1alpha1.OperatorConfiguration{}, "Routing"),
-			})
-			assert.Empty(t, diff)
-		})
-	}
-
-	for _, testCase := range expectedErrorTestCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).WithObjects(testCase.existedObjects...).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			_, _, err := devWorkspaceConfigReconciler.Reconcile(deployContext)
-			assert.Error(t, err)
-			assert.Regexp(t, regexp.MustCompile(testCase.expectedErrorMessage), err.Error(), "error message must match")
-		})
-	}
-}
-
-func TestReconcileDevWorkspaceConfigForContainerCapabilities(t *testing.T) {
-	type testCase struct {
-		name                   string
-		cheCluster             *chev2.CheCluster
-		existedObjects         []client.Object
-		expectedOperatorConfig *controllerv1alpha1.OperatorConfiguration
-	}
-
-	var unmasked = corev1.UnmaskedProcMount
-	var testCases = []testCase{
-		{
-			name: "Create DevWorkspaceOperatorConfig without Container Security Context if all capabilities disabled",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						DisableContainerRunCapabilities:   ptr.To(true),
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{},
-			},
-		},
-		{
-			name: "Create DevWorkspaceOperatorConfig with Container Security Context if container build enabled",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(false),
-						DisableContainerRunCapabilities:   ptr.To(true),
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ContainerSecurityContext: &corev1.SecurityContext{
-						Capabilities: &corev1.Capabilities{
-							Add: []corev1.Capability{
-								"SETGID",
-								"SETUID",
-							},
-						},
-						AllowPrivilegeEscalation: ptr.To(true),
-					},
-				},
-			},
-		},
-		{
-			name: "Create DevWorkspaceOperatorConfig with Container Security Context if container run enabled",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						WorkspacesPodAnnotations: map[string]string{
-							"annotation_1": "value_1",
-							"annotation_2": "value_1",
-						},
-						DisableContainerBuildCapabilities: ptr.To(true),
-						DisableContainerRunCapabilities:   ptr.To(false),
-						ContainerRunConfiguration: &chev2.ContainerRunConfiguration{
-							OpenShiftSecurityContextConstraint: "container-run",
-							WorkspacesPodAnnotations: map[string]string{
-								"annotation_1": "value_2",
-								"annotation_3": "value_2",
-							},
-							ContainerSecurityContext: &corev1.SecurityContext{
-								ProcMount: &unmasked,
-								Capabilities: &corev1.Capabilities{
-									Add: []corev1.Capability{"SETUID", "SETGID"},
-								},
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					HostUsers: ptr.To(false),
-					PodAnnotations: map[string]string{
-						"annotation_1": "value_2",
-						"annotation_2": "value_1",
-						"annotation_3": "value_2",
-					},
-					ContainerSecurityContext: &corev1.SecurityContext{
-						ProcMount: &unmasked,
-						Capabilities: &corev1.Capabilities{
-							Add: []corev1.Capability{"SETUID", "SETGID"},
-						},
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig by adding Container Security Context",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(false),
-						DisableContainerRunCapabilities:   ptr.To(true),
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ContainerSecurityContext: &corev1.SecurityContext{
-						Capabilities: &corev1.Capabilities{
-							Add: []corev1.Capability{
-								"SETGID",
-								"SETUID",
-							},
-						},
-						AllowPrivilegeEscalation: ptr.To(true),
-					},
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig by removing Container Security Context",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						DisableContainerRunCapabilities:   ptr.To(true),
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							ContainerSecurityContext: &corev1.SecurityContext{Capabilities: &corev1.Capabilities{
-								Add: []corev1.Capability{
-									"SETGID",
-									"SETUID",
-								},
-							},
-								AllowPrivilegeEscalation: ptr.To(true),
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{},
-			},
-		},
-		{
-			name: "Create DevWorkspaceOperatorConfig with Container Security Context if all capabilities enabled",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						WorkspacesPodAnnotations: map[string]string{
-							"annotation_1": "value_1",
-							"annotation_2": "value_1",
-						},
-						DisableContainerBuildCapabilities: ptr.To(false),
-						DisableContainerRunCapabilities:   ptr.To(false),
-						ContainerRunConfiguration: &chev2.ContainerRunConfiguration{
-							OpenShiftSecurityContextConstraint: "container-run",
-							WorkspacesPodAnnotations: map[string]string{
-								"annotation_1": "value_2",
-								"annotation_3": "value_2",
-							},
-							ContainerSecurityContext: &corev1.SecurityContext{
-								ProcMount: &unmasked,
-								Capabilities: &corev1.Capabilities{
-									Add: []corev1.Capability{"SETUID", "SETGID"},
-								},
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					HostUsers: ptr.To(false),
-					PodAnnotations: map[string]string{
-						"annotation_1": "value_2",
-						"annotation_2": "value_1",
-						"annotation_3": "value_2",
-					},
-					ContainerSecurityContext: &corev1.SecurityContext{
-						ProcMount: &unmasked,
-						Capabilities: &corev1.Capabilities{
-							Add: []corev1.Capability{"SETUID", "SETGID"},
-						},
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).WithObjects(testCase.existedObjects...).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(context.TODO(), types.NamespacedName{Name: devWorkspaceConfigName, Namespace: testCase.cheCluster.Namespace}, dwoc)
-			assert.NoError(t, err)
-
-			diff := cmp.Diff(testCase.expectedOperatorConfig, dwoc.Config,
-				cmp.Options{
-					cmpopts.IgnoreFields(controllerv1alpha1.WorkspaceConfig{}, "ServiceAccount", "ProjectCloneConfig", "DeploymentStrategy", "DefaultStorageSize", "StorageClassName"),
-					cmpopts.IgnoreFields(controllerv1alpha1.RoutingConfig{}, "TLSCertificateConfigmapRef"),
-				})
-			assert.Empty(t, diff)
-		})
-	}
-}
-
-func TestReconcileDevWorkspaceConfigProgressTimeout(t *testing.T) {
-	type testCase struct {
-		name                   string
-		cheCluster             *chev2.CheCluster
-		existedObjects         []client.Object
-		expectedOperatorConfig *controllerv1alpha1.OperatorConfiguration
-	}
-
-	var testCases = []testCase{
-		{
-			name: "Create DevWorkspaceOperatorConfig with progressTimeout",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						StartTimeoutSeconds:               ptr.To(int32(600)),
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ProgressTimeout:    "600s",
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig by adding progressTimeout",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      "eclipse-che",
-					Namespace: "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						StartTimeoutSeconds:               ptr.To(int32(600)),
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ProgressTimeout: "600s",
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig by overwriting progressTimeout",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						StartTimeoutSeconds:               ptr.To(int32(420)),
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							ProgressTimeout: "1h30m",
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ProgressTimeout: "420s",
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig by removing progressTimeout",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							ProgressTimeout: "1h30m",
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).WithObjects(testCase.existedObjects...).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(context.TODO(), types.NamespacedName{Name: devWorkspaceConfigName, Namespace: testCase.cheCluster.Namespace}, dwoc)
-			assert.NoError(t, err)
-
-			diff := cmp.Diff(testCase.expectedOperatorConfig, dwoc.Config,
-				cmp.Options{
-					cmpopts.IgnoreFields(controllerv1alpha1.WorkspaceConfig{}, "ServiceAccount", "DefaultStorageSize", "StorageClassName", "ProjectCloneConfig", "DeploymentStrategy"),
-					cmpopts.IgnoreFields(controllerv1alpha1.RoutingConfig{}, "TLSCertificateConfigmapRef"),
-				})
-			assert.Empty(t, diff)
-		})
-	}
-}
-
-func TestReconcileServiceAccountConfig(t *testing.T) {
-	type testCase struct {
-		name                   string
-		cheCluster             *chev2.CheCluster
-		expectedOperatorConfig *controllerv1alpha1.OperatorConfiguration
-	}
-
-	var testCases = []testCase{
-		{
-			name: "Case #1",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						ServiceAccount: "service-account",
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ServiceAccount: &controllerv1alpha1.ServiceAccountConfig{
-						ServiceAccountName: "service-account",
-						DisableCreation:    ptr.To(false),
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Case #2",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DefaultNamespace: chev2.DefaultNamespace{
-							AutoProvision: ptr.To(false),
-						},
-						ServiceAccount: "service-account",
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ServiceAccount: &controllerv1alpha1.ServiceAccountConfig{
-						ServiceAccountName: "service-account",
-						DisableCreation:    ptr.To(true),
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Case #3",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ServiceAccount: &controllerv1alpha1.ServiceAccountConfig{
-						DisableCreation: ptr.To(false),
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Case #4",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DefaultNamespace: chev2.DefaultNamespace{
-							AutoProvision: ptr.To(false),
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ServiceAccount: &controllerv1alpha1.ServiceAccountConfig{
-						DisableCreation: ptr.To(false),
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(context.TODO(), types.NamespacedName{Name: devWorkspaceConfigName, Namespace: testCase.cheCluster.Namespace}, dwoc)
-			assert.NoError(t, err)
-
-			assert.Equal(t, testCase.expectedOperatorConfig.Workspace.ServiceAccount, dwoc.Config.Workspace.ServiceAccount)
-		})
-	}
-}
-
-func TestReconcileDevWorkspaceConfigPodSchedulerName(t *testing.T) {
-	type testCase struct {
-		name                   string
-		cheCluster             *chev2.CheCluster
-		existedObjects         []client.Object
-		expectedOperatorConfig *controllerv1alpha1.OperatorConfiguration
-	}
-
-	var testCases = []testCase{
-		{
-			name: "Create DevWorkspaceOperatorConfig with podSchedulerName",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						PodSchedulerName: "test-scheduler",
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					SchedulerName:      "test-scheduler",
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig when PodSchedulerName is added",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						PodSchedulerName: "test-scheduler",
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					SchedulerName:      "test-scheduler",
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig when PodSchedulerName is changed",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						PodSchedulerName: "test-scheduler",
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							SchedulerName: "previous-scheduler",
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					SchedulerName:      "test-scheduler",
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig when PodSchedulerName is removed",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						PodSchedulerName: "",
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							SchedulerName: "previous-scheduler",
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).WithObjects(testCase.existedObjects...).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(context.TODO(), types.NamespacedName{Name: devWorkspaceConfigName, Namespace: testCase.cheCluster.Namespace}, dwoc)
-			assert.NoError(t, err)
-			assert.Equal(t, testCase.expectedOperatorConfig.Workspace.SchedulerName, dwoc.Config.Workspace.SchedulerName)
-		})
-	}
-}
-
-func TestReconcileDevWorkspaceConfigRuntimeClassName(t *testing.T) {
-	type testCase struct {
-		name                   string
-		cheCluster             *chev2.CheCluster
-		existedObjects         []client.Object
-		expectedOperatorConfig *controllerv1alpha1.OperatorConfiguration
-	}
-
-	var testCases = []testCase{
-		{
-			name: "Create DevWorkspaceOperatorConfig with RuntimeClassName",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						RuntimeClassName: ptr.To("test-runtime-class"),
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					RuntimeClassName: ptr.To("test-runtime-class"),
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig when RuntimeClassName is added",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						RuntimeClassName: ptr.To("test-runtime-class"),
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					RuntimeClassName: ptr.To("test-runtime-class"),
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig when RuntimeClassName is changed",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						RuntimeClassName: ptr.To("test-runtime-class"),
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							RuntimeClassName: ptr.To("previous-runtime-class"),
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					RuntimeClassName: ptr.To("test-runtime-class"),
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig when RuntimeClassName is removed",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						RuntimeClassName: nil,
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							RuntimeClassName: ptr.To("previous-runtime-class"),
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).WithObjects(testCase.existedObjects...).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(context.TODO(), types.NamespacedName{Name: devWorkspaceConfigName, Namespace: testCase.cheCluster.Namespace}, dwoc)
-			assert.NoError(t, err)
-			assert.Equal(t, testCase.expectedOperatorConfig.Workspace.SchedulerName, dwoc.Config.Workspace.SchedulerName)
-		})
-	}
-}
-
-func TestReconcileDevWorkspaceConfigServiceAccountTokens(t *testing.T) {
-	type testCase struct {
-		name                   string
-		cheCluster             *chev2.CheCluster
-		existedObjects         []client.Object
-		expectedOperatorConfig *controllerv1alpha1.OperatorConfiguration
-	}
-
-	var testCases = []testCase{
-		{
-			name: "Create DevWorkspaceOperatorConfig with ServiceAccountTokens",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						ServiceAccountTokens: []controllerv1alpha1.ServiceAccountToken{
-							{
-								Name:              "test-token-1",
-								MountPath:         "/var/run/secrets/tokens",
-								Audience:          "openshift",
-								ExpirationSeconds: 3600,
-							},
-							{
-								Name:              "test-token-2",
-								MountPath:         "/var/run/secrets/tokens",
-								Audience:          "kubernetes",
-								ExpirationSeconds: 180,
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ServiceAccount: &controllerv1alpha1.ServiceAccountConfig{
-						ServiceAccountTokens: []controllerv1alpha1.ServiceAccountToken{
-							{
-								Name:              "test-token-1",
-								MountPath:         "/var/run/secrets/tokens",
-								Audience:          "openshift",
-								ExpirationSeconds: 3600,
-							},
-							{
-								Name:              "test-token-2",
-								MountPath:         "/var/run/secrets/tokens",
-								Audience:          "kubernetes",
-								ExpirationSeconds: 180,
-							},
-						}},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig when ServiceAccountTokens are added",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						ServiceAccountTokens: []controllerv1alpha1.ServiceAccountToken{
-							{
-								Name:              "test-token",
-								MountPath:         "/var/run/secrets/tokens",
-								Audience:          "openshift",
-								ExpirationSeconds: 3600,
-							},
-							{
-								Name:              "test-token-2",
-								MountPath:         "/var/run/secrets/tokens",
-								Audience:          "kubernetes",
-								ExpirationSeconds: 180,
-							},
-						},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ServiceAccount: &controllerv1alpha1.ServiceAccountConfig{
-						ServiceAccountTokens: []controllerv1alpha1.ServiceAccountToken{
-							{
-								Name:              "test-token",
-								MountPath:         "/var/run/secrets/tokens",
-								Audience:          "openshift",
-								ExpirationSeconds: 3600,
-							},
-							{
-								Name:              "test-token-2",
-								MountPath:         "/var/run/secrets/tokens",
-								Audience:          "kubernetes",
-								ExpirationSeconds: 180,
-							},
-						},
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig when ServiceAccountTokens are changed",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						ServiceAccountTokens: []controllerv1alpha1.ServiceAccountToken{
-							{
-								Name:              "new-token",
-								MountPath:         "/var/run/secrets/tokens",
-								Audience:          "openshift",
-								ExpirationSeconds: 3600,
-							},
-						},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							ServiceAccount: &controllerv1alpha1.ServiceAccountConfig{
-								ServiceAccountTokens: []controllerv1alpha1.ServiceAccountToken{
-									{
-										Name:              "old-token",
-										MountPath:         "/var/run/secrets/tokens",
-										Audience:          "openshift",
-										ExpirationSeconds: 60,
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ServiceAccount: &controllerv1alpha1.ServiceAccountConfig{
-						ServiceAccountTokens: []controllerv1alpha1.ServiceAccountToken{
-							{
-								Name:              "new-token",
-								MountPath:         "/var/run/secrets/tokens",
-								Audience:          "openshift",
-								ExpirationSeconds: 3600,
-							},
-						},
-					},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig when ServiceAccountTokens are removed",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						ServiceAccountTokens: []controllerv1alpha1.ServiceAccountToken{},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							ServiceAccount: &controllerv1alpha1.ServiceAccountConfig{
-								ServiceAccountTokens: []controllerv1alpha1.ServiceAccountToken{
-									{
-										Name:              "test-token",
-										MountPath:         "/var/run/secrets/tokens",
-										Audience:          "openshift",
-										ExpirationSeconds: 3600,
-									},
-									{
-										Name:              "test-token-2",
-										MountPath:         "/var/run/secrets/tokens",
-										Audience:          "kubernetes",
-										ExpirationSeconds: 180,
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ServiceAccount:     &controllerv1alpha1.ServiceAccountConfig{},
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).WithObjects(testCase.existedObjects...).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(context.TODO(), types.NamespacedName{Name: devWorkspaceConfigName, Namespace: testCase.cheCluster.Namespace}, dwoc)
-			assert.NoError(t, err)
-			assert.Equal(t, testCase.expectedOperatorConfig.Workspace.ServiceAccount.ServiceAccountTokens, dwoc.Config.Workspace.ServiceAccount.ServiceAccountTokens)
-		})
-	}
-}
-
-func TestReconcileDevWorkspaceConfigDeploymentStrategy(t *testing.T) {
-	type testCase struct {
-		name                   string
-		cheCluster             *chev2.CheCluster
-		existedObjects         []client.Object
-		expectedOperatorConfig *controllerv1alpha1.OperatorConfiguration
-	}
-
-	var testCases = []testCase{
-		{
-			name: "Create DevWorkspaceOperatorConfig with DeploymentStrategy",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DeploymentStrategy: "Recreate",
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig when DeploymentStrategy is added",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DeploymentStrategy: "Recreate",
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					DeploymentStrategy: "Recreate",
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig when DeploymentStrategy is changed",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DeploymentStrategy: "RollingUpdate",
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							DeploymentStrategy: "Recreate",
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					DeploymentStrategy: "RollingUpdate",
-				},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).WithObjects(testCase.existedObjects...).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(context.TODO(), types.NamespacedName{Name: devWorkspaceConfigName, Namespace: testCase.cheCluster.Namespace}, dwoc)
-			assert.NoError(t, err)
-			assert.Equal(t, testCase.expectedOperatorConfig.Workspace.DeploymentStrategy, dwoc.Config.Workspace.DeploymentStrategy)
-		})
-	}
-}
-
-func TestReconcileDevWorkspaceProjectCloneConfig(t *testing.T) {
-	const testNamespace = "eclipse-che"
-	testMemLimit := resource.MustParse("2Gi")
-	testCpuLimit := resource.MustParse("1000m")
-	testMemRequest := resource.MustParse("1Gi")
-	testCpuRequest := resource.MustParse("500m")
-
-	type testCase struct {
-		name                       string
-		cheProjectCloneConfig      *chev2.Container
-		expectedDevWorkspaceConfig *controllerv1alpha1.ProjectCloneConfig
-		existingDevWorkspaceConfig *controllerv1alpha1.ProjectCloneConfig
-	}
-
-	tests := []testCase{
-		{
-			name: "Syncs Che project clone config to DevWorkspaceOperatorConfig",
-			cheProjectCloneConfig: &chev2.Container{
-				Name:            "project-clone",
-				Image:           "test-image",
-				ImagePullPolicy: "IfNotPresent",
-				Env: []corev1.EnvVar{
-					{Name: "test-env-1", Value: "test-val-1"},
-					{Name: "test-env-2", Value: "test-val-2"},
-				},
-				Resources: &chev2.ResourceRequirements{
-					Limits: &chev2.ResourceList{
-						Memory: &testMemLimit,
-						Cpu:    &testCpuLimit,
-					},
-					Requests: &chev2.ResourceList{
-						Memory: &testMemRequest,
-						Cpu:    &testCpuRequest,
-					},
-				},
-			},
-			expectedDevWorkspaceConfig: &controllerv1alpha1.ProjectCloneConfig{
-				Image:           "test-image",
-				ImagePullPolicy: "IfNotPresent",
-				Env: []corev1.EnvVar{
-					{Name: "test-env-1", Value: "test-val-1"},
-					{Name: "test-env-2", Value: "test-val-2"},
-				},
-				Resources: &corev1.ResourceRequirements{
-					Limits: corev1.ResourceList{
-						corev1.ResourceMemory: testMemLimit,
-						corev1.ResourceCPU:    testCpuLimit,
-					},
-					Requests: corev1.ResourceList{
-						corev1.ResourceMemory: testMemRequest,
-						corev1.ResourceCPU:    testCpuRequest,
-					},
-				},
-			},
-		},
-		{
-			name: "Updates existing DevWorkspaceOperatorConfig with new Che project clone config",
-			cheProjectCloneConfig: &chev2.Container{
-				Name:            "project-clone",
-				Image:           "test-image",
-				ImagePullPolicy: "IfNotPresent",
-				Env: []corev1.EnvVar{
-					{Name: "test-env-1", Value: "test-val-1"},
-					{Name: "test-env-2", Value: "test-val-2"},
-				},
-				Resources: &chev2.ResourceRequirements{
-					Limits: &chev2.ResourceList{
-						Memory: &testMemLimit,
-						Cpu:    &testCpuLimit,
-					},
-					Requests: &chev2.ResourceList{
-						Memory: &testMemRequest,
-						Cpu:    &testCpuRequest,
-					},
-				},
-			},
-			expectedDevWorkspaceConfig: &controllerv1alpha1.ProjectCloneConfig{
-				Image:           "test-image",
-				ImagePullPolicy: "IfNotPresent",
-				Env: []corev1.EnvVar{
-					{Name: "test-env-1", Value: "test-val-1"},
-					{Name: "test-env-2", Value: "test-val-2"},
-				},
-				Resources: &corev1.ResourceRequirements{
-					Limits: corev1.ResourceList{
-						corev1.ResourceMemory: testMemLimit,
-						corev1.ResourceCPU:    testCpuLimit,
-					},
-					Requests: corev1.ResourceList{
-						corev1.ResourceMemory: testMemRequest,
-						corev1.ResourceCPU:    testCpuRequest,
-					},
-				},
-			},
-			existingDevWorkspaceConfig: &controllerv1alpha1.ProjectCloneConfig{
-				Image:           "other image",
-				ImagePullPolicy: "Always",
-				Env: []corev1.EnvVar{
-					{Name: "other-env", Value: "other-val"},
-				},
-				Resources: &corev1.ResourceRequirements{
-					Limits: corev1.ResourceList{
-						corev1.ResourceMemory: resource.MustParse("1234Mi"),
-						corev1.ResourceCPU:    resource.MustParse("1234m"),
-					},
-					Requests: corev1.ResourceList{
-						corev1.ResourceMemory: resource.MustParse("1111Mi"),
-						corev1.ResourceCPU:    resource.MustParse("1111m"),
-					},
-				},
-			},
-		},
-		{
-			name: "Removes fields from existing config when removed from CheCluster",
-			cheProjectCloneConfig: &chev2.Container{
-				Name:            "",
-				Image:           "",
-				ImagePullPolicy: "",
-				Env:             nil,
-				Resources:       nil,
-			},
-			expectedDevWorkspaceConfig: &controllerv1alpha1.ProjectCloneConfig{
-				Image:           "",
-				ImagePullPolicy: "",
-				Env:             nil,
-				Resources:       nil,
-			},
-			existingDevWorkspaceConfig: &controllerv1alpha1.ProjectCloneConfig{
-				Image:           "other image",
-				ImagePullPolicy: "Always",
-				Env: []corev1.EnvVar{
-					{Name: "other-env", Value: "other-val"},
-				},
-				Resources: &corev1.ResourceRequirements{
-					Limits: corev1.ResourceList{
-						corev1.ResourceMemory: resource.MustParse("1234Mi"),
-						corev1.ResourceCPU:    resource.MustParse("1234m"),
-					},
-					Requests: corev1.ResourceList{
-						corev1.ResourceMemory: resource.MustParse("1111Mi"),
-						corev1.ResourceCPU:    resource.MustParse("1111m"),
-					},
-				},
-			},
-		},
-	}
-
-	for _, testCase := range tests {
-		t.Run(testCase.name, func(t *testing.T) {
-			cheCluster := &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: testNamespace,
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						ProjectCloneContainer: testCase.cheProjectCloneConfig,
-					},
-				},
-			}
-			existingDWOC := &controllerv1alpha1.DevWorkspaceOperatorConfig{
-				ObjectMeta: metav1.ObjectMeta{
-					Name:      devWorkspaceConfigName,
-					Namespace: testNamespace,
-				},
-				Config: &controllerv1alpha1.OperatorConfiguration{
-					Workspace: &controllerv1alpha1.WorkspaceConfig{
-						ProjectCloneConfig: testCase.existingDevWorkspaceConfig,
-					},
-				},
-			}
-			runtimeDWOC := client.Object(existingDWOC)
-
-			deployContext := test.NewCtxBuilder().WithCheCluster(cheCluster).WithObjects(runtimeDWOC).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(context.TODO(), types.NamespacedName{Name: devWorkspaceConfigName, Namespace: testNamespace}, dwoc)
-			assert.NoError(t, err)
-
-			diff := cmp.Diff(testCase.expectedDevWorkspaceConfig, dwoc.Config.Workspace.ProjectCloneConfig)
-			assert.Empty(t, diff)
-		})
-	}
-
-}
-
-func TestReconcileDevWorkspaceConfigPersistUserHome(t *testing.T) {
-	type testCase struct {
-		name                   string
-		cheCluster             *chev2.CheCluster
-		existedObjects         []client.Object
-		expectedOperatorConfig *controllerv1alpha1.OperatorConfiguration
-	}
-
-	var testCases = []testCase{
-		{
-			name: "Create DevWorkspaceOperatorConfig when PersistUserHome is enabled",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						PersistUserHome: &chev2.PersistentHomeConfig{
-							Enabled: ptr.To(true),
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					PersistUserHome: &controllerv1alpha1.PersistentHomeConfig{
-						Enabled: ptr.To(true),
-					},
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig that does not have PersistUserHome config defined, when PersistUserHome is enabled",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						PersistUserHome: &chev2.PersistentHomeConfig{
-							Enabled: ptr.To(true),
-						},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					PersistUserHome: &controllerv1alpha1.PersistentHomeConfig{
-						Enabled: ptr.To(true),
-					},
-				},
-			},
-		},
-		{
-			name: "Set DevWorkspaceOperatorConfig PersistUserHome enabled when PersistUserHome is enabled",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						PersistUserHome: &chev2.PersistentHomeConfig{
-							Enabled: ptr.To(true),
-						},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							PersistUserHome: &controllerv1alpha1.PersistentHomeConfig{
-								Enabled: ptr.To(false),
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					PersistUserHome: &controllerv1alpha1.PersistentHomeConfig{
-						Enabled: ptr.To(true),
-					},
-				},
-			},
-		},
-		{
-			name: "Set DevWorkspaceOperatorConfig PersistUserHome disableInitContainer to true",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						PersistUserHome: &chev2.PersistentHomeConfig{
-							Enabled:              ptr.To(true),
-							DisableInitContainer: ptr.To(true),
-						},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							PersistUserHome: &controllerv1alpha1.PersistentHomeConfig{
-								Enabled: ptr.To(false),
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					PersistUserHome: &controllerv1alpha1.PersistentHomeConfig{
-						Enabled:              ptr.To(true),
-						DisableInitContainer: ptr.To(true),
-					},
-				},
-			},
-		},
-		{
-			name: "Set DevWorkspaceOperatorConfig PersistUserHome disableInitContainer to false when initially true",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						PersistUserHome: &chev2.PersistentHomeConfig{
-							Enabled:              ptr.To(true),
-							DisableInitContainer: ptr.To(false),
-						},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							PersistUserHome: &controllerv1alpha1.PersistentHomeConfig{
-								Enabled:              ptr.To(true),
-								DisableInitContainer: ptr.To(true),
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					PersistUserHome: &controllerv1alpha1.PersistentHomeConfig{
-						Enabled:              ptr.To(true),
-						DisableInitContainer: ptr.To(false),
-					},
-				},
-			},
-		},
-		{
-			name: "Set DevWorkspaceOperatorConfig PersistUserHome disabled when PersistUserHome is disabled",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						PersistUserHome: &chev2.PersistentHomeConfig{
-							Enabled: ptr.To(false),
-						},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							PersistUserHome: &controllerv1alpha1.PersistentHomeConfig{
-								Enabled: ptr.To(true),
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					PersistUserHome: &controllerv1alpha1.PersistentHomeConfig{
-						Enabled: ptr.To(false),
-					},
-				},
-			},
-		},
-		{
-			name: "Remove PersistUserHome config from existing DevWorkspaceOperatorConfig when PersistUserHome is removed from Che Cluster",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							PersistUserHome: &controllerv1alpha1.PersistentHomeConfig{
-								Enabled: ptr.To(true),
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).WithObjects(testCase.existedObjects...).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(context.TODO(), types.NamespacedName{Name: devWorkspaceConfigName, Namespace: testCase.cheCluster.Namespace}, dwoc)
-			assert.NoError(t, err)
-			diff := cmp.Diff(testCase.expectedOperatorConfig, dwoc.Config, cmp.Options{
-				cmpopts.IgnoreFields(controllerv1alpha1.WorkspaceConfig{}, "ServiceAccount", "DeploymentStrategy", "ContainerSecurityContext"),
-				cmpopts.IgnoreFields(controllerv1alpha1.RoutingConfig{}, "TLSCertificateConfigmapRef"),
-			})
-			assert.Empty(t, diff)
-		})
-	}
-}
-
-func TestReconcileDevWorkspaceContainerSecurityContext(t *testing.T) {
-	type testCase struct {
-		name                   string
-		cheCluster             *chev2.CheCluster
-		existedObjects         []client.Object
-		expectedOperatorConfig *controllerv1alpha1.OperatorConfiguration
-	}
-
-	var testCases = []testCase{
-		{
-			name: "Create DevWorkspaceOperatorConfig with Container security context",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						// We disable container build capabilities so that it does not override the container security context we configured
-						DisableContainerBuildCapabilities: ptr.To(true),
-						DisableContainerRunCapabilities:   ptr.To(true),
-						Security: chev2.WorkspaceSecurityConfig{
-							ContainerSecurityContext: &corev1.SecurityContext{
-								Capabilities: &corev1.Capabilities{
-									Add: []corev1.Capability{
-										"SYS_TIME",
-										"SETGID",
-										"SETUID",
-									},
-									Drop: []corev1.Capability{
-										"CHOWN",
-										"KILL",
-									},
-								},
-								AllowPrivilegeEscalation: ptr.To(false),
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ContainerSecurityContext: &corev1.SecurityContext{
-						Capabilities: &corev1.Capabilities{
-							Add: []corev1.Capability{
-								"SYS_TIME",
-								"SETGID",
-								"SETUID",
-							},
-							Drop: []corev1.Capability{
-								"CHOWN",
-								"KILL",
-							},
-						},
-						AllowPrivilegeEscalation: ptr.To(false),
-					},
-				},
-			},
-		},
-		{
-			name: "Updates existing DevWorkspaceOperatorConfig when Container security context is added",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						DisableContainerRunCapabilities:   ptr.To(true),
-						Security: chev2.WorkspaceSecurityConfig{
-							ContainerSecurityContext: &corev1.SecurityContext{
-								Capabilities: &corev1.Capabilities{
-									Add: []corev1.Capability{
-										"SYS_TIME",
-										"SETGID",
-										"SETUID",
-									},
-									Drop: []corev1.Capability{
-										"CHOWN",
-										"KILL",
-									},
-								},
-								AllowPrivilegeEscalation: ptr.To(false),
-							}},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ContainerSecurityContext: &corev1.SecurityContext{
-						Capabilities: &corev1.Capabilities{
-							Add: []corev1.Capability{
-								"SYS_TIME",
-								"SETGID",
-								"SETUID",
-							},
-							Drop: []corev1.Capability{
-								"CHOWN",
-								"KILL",
-							},
-						},
-						AllowPrivilegeEscalation: ptr.To(false),
-					},
-				},
-			},
-		},
-		{
-			name: "Updates existing DevWorkspaceOperatorConfig when Container security is changed",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						DisableContainerRunCapabilities:   ptr.To(true),
-						Security: chev2.WorkspaceSecurityConfig{
-							ContainerSecurityContext: &corev1.SecurityContext{
-								Capabilities: &corev1.Capabilities{
-									Add: []corev1.Capability{
-										"SYS_TIME",
-										"SETGID",
-										"SETUID",
-									},
-									Drop: []corev1.Capability{
-										"CHOWN",
-										"KILL",
-									},
-								},
-								AllowPrivilegeEscalation: ptr.To(false),
-							},
-						},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							ContainerSecurityContext: &corev1.SecurityContext{
-								Capabilities: &corev1.Capabilities{
-									Add: []corev1.Capability{
-										"KILL",
-									},
-									Drop: []corev1.Capability{
-										"SYS_TIME",
-									},
-								},
-								AllowPrivilegeEscalation: ptr.To(true),
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ContainerSecurityContext: &corev1.SecurityContext{
-						Capabilities: &corev1.Capabilities{
-							Add: []corev1.Capability{
-								"SYS_TIME",
-								"SETGID",
-								"SETUID",
-							},
-							Drop: []corev1.Capability{
-								"CHOWN",
-								"KILL",
-							},
-						},
-						AllowPrivilegeEscalation: ptr.To(false),
-					},
-				},
-			},
-		},
-		{
-			name: "Updates existing DevWorkspaceOperatorConfig when Container security is removed",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						DisableContainerBuildCapabilities: ptr.To(true),
-						DisableContainerRunCapabilities:   ptr.To(true),
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							ContainerSecurityContext: &corev1.SecurityContext{
-								Capabilities: &corev1.Capabilities{
-									Add: []corev1.Capability{
-										"KILL",
-									},
-									Drop: []corev1.Capability{
-										"SYS_TIME",
-									},
-								},
-								AllowPrivilegeEscalation: ptr.To(true),
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(context.TODO(), types.NamespacedName{Name: devWorkspaceConfigName, Namespace: testCase.cheCluster.Namespace}, dwoc)
-			assert.NoError(t, err)
-
-			sortCapabilities := func(capabilities []corev1.Capability) func(i, j int) bool {
-				return func(i, j int) bool {
-					return capabilities[i] > capabilities[j]
-				}
-			}
-			expectedContainerSecurityContext := testCase.expectedOperatorConfig.Workspace.ContainerSecurityContext
-			actualContainerSecurityContext := dwoc.Config.Workspace.ContainerSecurityContext
-			if expectedContainerSecurityContext != nil {
-				sort.Slice(expectedContainerSecurityContext.Capabilities.Add, sortCapabilities(expectedContainerSecurityContext.Capabilities.Add))
-				sort.Slice(expectedContainerSecurityContext.Capabilities.Drop, sortCapabilities(expectedContainerSecurityContext.Capabilities.Drop))
-			}
-			if actualContainerSecurityContext != nil {
-				sort.Slice(actualContainerSecurityContext.Capabilities.Add, sortCapabilities(actualContainerSecurityContext.Capabilities.Add))
-				sort.Slice(actualContainerSecurityContext.Capabilities.Drop, sortCapabilities(actualContainerSecurityContext.Capabilities.Drop))
-
-			}
-
-			assert.Equal(t, expectedContainerSecurityContext, actualContainerSecurityContext,
-				fmt.Sprintf("Did not get expected ContainerSecurityContext.\nDiff:%s", cmp.Diff(expectedContainerSecurityContext, actualContainerSecurityContext)))
-		})
-	}
-}
-
-func TestReconcileDevWorkspacePodSecurityContext(t *testing.T) {
-	type testCase struct {
-		name                   string
-		cheCluster             *chev2.CheCluster
-		existedObjects         []client.Object
-		expectedOperatorConfig *controllerv1alpha1.OperatorConfiguration
-	}
-
-	configuredPodSecurityContext := &corev1.PodSecurityContext{
-		RunAsUser:    ptr.To(int64(0)),
-		RunAsGroup:   ptr.To(int64(0)),
-		RunAsNonRoot: ptr.To(false),
-	}
-
-	var testCases = []testCase{
-		{
-			name: "Create DevWorkspaceOperatorConfig with Pod security context",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						Security: chev2.WorkspaceSecurityConfig{
-							PodSecurityContext: configuredPodSecurityContext,
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					PodSecurityContext: configuredPodSecurityContext,
-				},
-			},
-		},
-		{
-			name: "Updates existing DevWorkspaceOperatorConfig with Pod security context",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						Security: chev2.WorkspaceSecurityConfig{
-							PodSecurityContext: configuredPodSecurityContext,
-						},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					PodSecurityContext: configuredPodSecurityContext,
-				},
-			},
-		},
-		{
-			name: "Updates existing DevWorkspaceOperatorConfig when Pod security context is changed",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						Security: chev2.WorkspaceSecurityConfig{
-							PodSecurityContext: configuredPodSecurityContext,
-						},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							PodSecurityContext: &corev1.PodSecurityContext{
-								RunAsUser:    ptr.To(int64(1000)),
-								RunAsGroup:   ptr.To(int64(10001)),
-								RunAsNonRoot: ptr.To(true),
-								SupplementalGroups: []int64{
-									5,
-								},
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					PodSecurityContext: configuredPodSecurityContext,
-				},
-			},
-		},
-		{
-			name: "Updates existing DevWorkspaceOperatorConfig when Pod security context is removed",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							PodSecurityContext: &corev1.PodSecurityContext{
-								RunAsUser:    ptr.To(int64(1000)),
-								RunAsGroup:   ptr.To(int64(10001)),
-								RunAsNonRoot: ptr.To(true),
-								SupplementalGroups: []int64{
-									5,
-								},
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(context.TODO(), types.NamespacedName{Name: devWorkspaceConfigName, Namespace: testCase.cheCluster.Namespace}, dwoc)
-			assert.NoError(t, err)
-			assert.Equal(t, testCase.expectedOperatorConfig.Workspace.PodSecurityContext, dwoc.Config.Workspace.PodSecurityContext,
-				fmt.Sprintf("Did not get expected PodSecurityContext.\nDiff:%s", cmp.Diff(testCase.expectedOperatorConfig.Workspace.PodSecurityContext, dwoc.Config.Workspace.PodSecurityContext)))
-		})
-	}
-}
-
-func TestReconcileDevWorkspaceImagePullPolicy(t *testing.T) {
-	type testCase struct {
-		name                   string
-		cheCluster             *chev2.CheCluster
-		existedObjects         []client.Object
-		expectedOperatorConfig *controllerv1alpha1.OperatorConfiguration
-	}
-
-	var testCases = []testCase{
-		{
-			name: "Set specific pull policy",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						ImagePullPolicy: corev1.PullAlways,
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ImagePullPolicy: string(corev1.PullAlways),
-				},
-			},
-		},
-		{
-			name: "Clean up pull policy",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						ImagePullPolicy: "",
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					ImagePullPolicy: "",
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							ImagePullPolicy: string(corev1.PullAlways),
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(context.TODO(), types.NamespacedName{Name: devWorkspaceConfigName, Namespace: testCase.cheCluster.Namespace}, dwoc)
-			assert.NoError(t, err)
-			assert.Equal(t, testCase.expectedOperatorConfig.Workspace.ImagePullPolicy, dwoc.Config.Workspace.ImagePullPolicy)
-		})
-	}
-}
-
-func TestReconcileDevWorkspaceAnnotations(t *testing.T) {
-	type testCase struct {
-		name                   string
-		cheCluster             *chev2.CheCluster
-		existedObjects         []client.Object
-		expectedOperatorConfig *controllerv1alpha1.OperatorConfiguration
-	}
-
-	var testCases = []testCase{
-		{
-			name: "Set specific annotations",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						WorkspacesPodAnnotations: map[string]string{
-							"key1": "value1",
-							"key2": "value2",
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					PodAnnotations: map[string]string{
-						"key1": "value1",
-						"key2": "value2",
-					},
-				},
-			},
-		},
-		{
-			name: "Remove annotations",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						WorkspacesPodAnnotations: nil,
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					PodAnnotations: nil,
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							PodAnnotations: map[string]string{
-								"key1": "value1",
-								"key2": "value2",
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(context.TODO(), types.NamespacedName{Name: devWorkspaceConfigName, Namespace: testCase.cheCluster.Namespace}, dwoc)
-			assert.NoError(t, err)
-			assert.Equal(t, testCase.expectedOperatorConfig.Workspace.PodAnnotations, dwoc.Config.Workspace.PodAnnotations)
-		})
-	}
-}
-
-func TestReconcileDevWorkspaceIgnoredUnrecoverableEvents(t *testing.T) {
-	type testCase struct {
-		name                   string
-		cheCluster             *chev2.CheCluster
-		existedObjects         []client.Object
-		expectedOperatorConfig *controllerv1alpha1.OperatorConfiguration
-	}
-
-	var testCases = []testCase{
-		{
-			name: "Set events",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						IgnoredUnrecoverableEvents: []string{
-							"value1",
-							"value2",
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					IgnoredUnrecoverableEvents: []string{
-						"value1",
-						"value2",
-					},
-				},
-			},
-		},
-		{
-			name: "Remove events",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						IgnoredUnrecoverableEvents: nil,
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					IgnoredUnrecoverableEvents: nil,
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							IgnoredUnrecoverableEvents: []string{
-								"value1",
-								"value2",
-							},
-						},
-					},
-				},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(context.TODO(), types.NamespacedName{Name: devWorkspaceConfigName, Namespace: testCase.cheCluster.Namespace}, dwoc)
-			assert.NoError(t, err)
-			assert.Equal(t, testCase.expectedOperatorConfig.Workspace.IgnoredUnrecoverableEvents, dwoc.Config.Workspace.IgnoredUnrecoverableEvents)
-		})
-	}
-}
-
-func TestReconcileDevWorkspaceConfigForInitContainers(t *testing.T) {
-	type testCase struct {
-		name                   string
-		cheCluster             *chev2.CheCluster
-		existedObjects         []client.Object
-		expectedOperatorConfig *controllerv1alpha1.OperatorConfiguration
-	}
-
-	testCases := []testCase{
-		{
-			name: "Create DevWorkspaceOperatorConfig with InitContainers",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						InitContainers: []corev1.Container{
-							{
-								Name:  "init-container-1",
-								Image: "init-image:latest",
-								Command: []string{
-									"/bin/sh",
-									"-c",
-									"echo 'Initializing workspace'",
-								},
-							},
-							{
-								Name:  "init-container-2",
-								Image: "init-image-2:v1.0",
-								Env: []corev1.EnvVar{
-									{
-										Name:  "INIT_VAR",
-										Value: "init-value",
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					InitContainers: []corev1.Container{
-						{
-							Name:  "init-container-1",
-							Image: "init-image:latest",
-							Command: []string{
-								"/bin/sh",
-								"-c",
-								"echo 'Initializing workspace'",
-							},
-						},
-						{
-							Name:  "init-container-2",
-							Image: "init-image-2:v1.0",
-							Env: []corev1.EnvVar{
-								{
-									Name:  "INIT_VAR",
-									Value: "init-value",
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "Create DevWorkspaceOperatorConfig without InitContainers",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{},
-			},
-		},
-		{
-			name: "Update DevWorkspaceOperatorConfig with InitContainers",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{
-						InitContainers: []corev1.Container{
-							{
-								Name:  "new-init-container",
-								Image: "new-init:v2.0",
-								VolumeMounts: []corev1.VolumeMount{
-									{
-										Name:      "config-volume",
-										MountPath: "/etc/config",
-									},
-								},
-							},
-						},
-					},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							InitContainers: []corev1.Container{
-								{
-									Name:  "old-init-container",
-									Image: "old-init:v1.0",
-								},
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{
-					InitContainers: []corev1.Container{
-						{
-							Name:  "new-init-container",
-							Image: "new-init:v2.0",
-							VolumeMounts: []corev1.VolumeMount{
-								{
-									Name:      "config-volume",
-									MountPath: "/etc/config",
-								},
-							},
-						},
-					},
-				},
-			},
-		},
-		{
-			name: "Clear InitContainers from DevWorkspaceOperatorConfig",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-				Spec: chev2.CheClusterSpec{
-					DevEnvironments: chev2.CheClusterDevEnvironments{},
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Workspace: &controllerv1alpha1.WorkspaceConfig{
-							InitContainers: []corev1.Container{
-								{
-									Name:  "init-to-remove",
-									Image: "init:v1.0",
-								},
-							},
-						},
-					},
-				},
-			},
-			expectedOperatorConfig: &controllerv1alpha1.OperatorConfiguration{
-				Workspace: &controllerv1alpha1.WorkspaceConfig{},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).WithObjects(testCase.existedObjects...).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(context.TODO(), types.NamespacedName{Name: devWorkspaceConfigName, Namespace: testCase.cheCluster.Namespace}, dwoc)
-			assert.NoError(t, err)
-
-			assert.Equal(t, testCase.expectedOperatorConfig.Workspace.InitContainers, dwoc.Config.Workspace.InitContainers)
-		})
-	}
-}
-
-func TestReconcileDevWorkspaceConfigTLSCertificateConfigmapRef(t *testing.T) {
-	type testCase struct {
-		name                  string
-		cheCluster            *chev2.CheCluster
-		existedObjects        []client.Object
-		expectedRoutingConfig *controllerv1alpha1.RoutingConfig
-	}
-
-	var testCases = []testCase{
-		{
-			name: "Create DevWorkspaceOperatorConfig with TLSCertificateConfigmapRef when CA bundle ConfigMap exists with data",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-			},
-			existedObjects: []client.Object{
-				&corev1.ConfigMap{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      tls.CheMergedCABundleCertsCMName,
-						Namespace: "eclipse-che",
-					},
-					Data: map[string]string{
-						"ca-bundle.crt": "certificate-data",
-					},
-				},
-			},
-			expectedRoutingConfig: &controllerv1alpha1.RoutingConfig{
-				TLSCertificateConfigmapRef: &controllerv1alpha1.ConfigmapReference{
-					Name:      tls.CheMergedCABundleCertsCMName,
-					Namespace: "eclipse-che",
-				},
-			},
-		},
-		{
-			name: "Update existing DevWorkspaceOperatorConfig with TLSCertificateConfigmapRef when CA bundle ConfigMap exists with data",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-			},
-			existedObjects: []client.Object{
-				&corev1.ConfigMap{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      tls.CheMergedCABundleCertsCMName,
-						Namespace: "eclipse-che",
-					},
-					Data: map[string]string{
-						"ca-bundle.crt": "certificate-data",
-					},
-				},
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Routing: &controllerv1alpha1.RoutingConfig{
-							DefaultRoutingClass: "che",
-						},
-					},
-				},
-			},
-			expectedRoutingConfig: &controllerv1alpha1.RoutingConfig{
-				DefaultRoutingClass: "che",
-				TLSCertificateConfigmapRef: &controllerv1alpha1.ConfigmapReference{
-					Name:      tls.CheMergedCABundleCertsCMName,
-					Namespace: "eclipse-che",
-				},
-			},
-		},
-		{
-			name: "Do not set TLSCertificateConfigmapRef when CA bundle ConfigMap does not exist",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-			},
-			expectedRoutingConfig: nil,
-		},
-		{
-			name: "Do not set TLSCertificateConfigmapRef when CA bundle ConfigMap exists but has no data",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-			},
-			existedObjects: []client.Object{
-				&corev1.ConfigMap{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      tls.CheMergedCABundleCertsCMName,
-						Namespace: "eclipse-che",
-					},
-					Data: map[string]string{},
-				},
-			},
-			expectedRoutingConfig: nil,
-		},
-		{
-			name: "Clear TLSCertificateConfigmapRef when CA bundle ConfigMap is removed",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Routing: &controllerv1alpha1.RoutingConfig{
-							TLSCertificateConfigmapRef: &controllerv1alpha1.ConfigmapReference{
-								Name:      tls.CheMergedCABundleCertsCMName,
-								Namespace: "eclipse-che",
-							},
-						},
-					},
-				},
-			},
-			expectedRoutingConfig: nil,
-		},
-		{
-			name: "Clear TLSCertificateConfigmapRef when CA bundle ConfigMap data is emptied",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-			},
-			existedObjects: []client.Object{
-				&corev1.ConfigMap{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      tls.CheMergedCABundleCertsCMName,
-						Namespace: "eclipse-che",
-					},
-					Data: map[string]string{},
-				},
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Routing: &controllerv1alpha1.RoutingConfig{
-							TLSCertificateConfigmapRef: &controllerv1alpha1.ConfigmapReference{
-								Name:      tls.CheMergedCABundleCertsCMName,
-								Namespace: "eclipse-che",
-							},
-						},
-					},
-				},
-			},
-			expectedRoutingConfig: nil,
-		},
-		{
-			name: "Re-reconcile is stable when TLSCertificateConfigmapRef already matches",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-			},
-			existedObjects: []client.Object{
-				&corev1.ConfigMap{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      tls.CheMergedCABundleCertsCMName,
-						Namespace: "eclipse-che",
-					},
-					Data: map[string]string{
-						"ca-bundle.crt": "certificate-data",
-					},
-				},
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Routing: &controllerv1alpha1.RoutingConfig{
-							TLSCertificateConfigmapRef: &controllerv1alpha1.ConfigmapReference{
-								Name:      tls.CheMergedCABundleCertsCMName,
-								Namespace: "eclipse-che",
-							},
-						},
-					},
-				},
-			},
-			expectedRoutingConfig: &controllerv1alpha1.RoutingConfig{
-				TLSCertificateConfigmapRef: &controllerv1alpha1.ConfigmapReference{
-					Name:      tls.CheMergedCABundleCertsCMName,
-					Namespace: "eclipse-che",
-				},
-			},
-		},
-		{
-			name: "Clear TLSCertificateConfigmapRef while preserving other routing config",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-			},
-			existedObjects: []client.Object{
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Routing: &controllerv1alpha1.RoutingConfig{
-							DefaultRoutingClass: "che",
-							TLSCertificateConfigmapRef: &controllerv1alpha1.ConfigmapReference{
-								Name:      tls.CheMergedCABundleCertsCMName,
-								Namespace: "eclipse-che",
-							},
-						},
-					},
-				},
-			},
-			expectedRoutingConfig: &controllerv1alpha1.RoutingConfig{
-				DefaultRoutingClass: "che",
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).WithObjects(testCase.existedObjects...).Build()
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(
-				context.TODO(),
-				types.NamespacedName{
-					Name:      devWorkspaceConfigName,
-					Namespace: testCase.cheCluster.Namespace,
-				},
-				dwoc,
-			)
-
-			assert.NoError(t, err)
-			assert.Equal(t, testCase.expectedRoutingConfig, dwoc.Config.Routing)
-		})
-	}
-}
-
-func TestReconcileDevWorkspaceConfigProxyAndTLSComposition(t *testing.T) {
-	type testCase struct {
-		name                  string
-		cheCluster            *chev2.CheCluster
-		existedObjects        []client.Object
-		httpProxy             string
-		httpsProxy            string
-		expectedRoutingConfig *controllerv1alpha1.RoutingConfig
-	}
-
-	var testCases = []testCase{
-		{
-			name: "Both proxy and TLS certificate configmap ref are set on RoutingConfig",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-			},
-			existedObjects: []client.Object{
-				&corev1.ConfigMap{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      tls.CheMergedCABundleCertsCMName,
-						Namespace: "eclipse-che",
-					},
-					Data: map[string]string{
-						"ca-bundle.crt": "certificate-data",
-					},
-				},
-			},
-			httpProxy:  "http://proxy.example.com:3128",
-			httpsProxy: "https://proxy.example.com:3128",
-			expectedRoutingConfig: &controllerv1alpha1.RoutingConfig{
-				TLSCertificateConfigmapRef: &controllerv1alpha1.ConfigmapReference{
-					Name:      tls.CheMergedCABundleCertsCMName,
-					Namespace: "eclipse-che",
-				},
-				ProxyConfig: &controllerv1alpha1.Proxy{
-					HttpProxy:  ptr.To(""),
-					HttpsProxy: ptr.To(""),
-					NoProxy:    ptr.To(""),
-				},
-			},
-		},
-		{
-			name: "Proxy set without TLS certificate configmap does not clobber routing",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-			},
-			httpProxy:  "http://proxy.example.com:3128",
-			httpsProxy: "https://proxy.example.com:3128",
-			expectedRoutingConfig: &controllerv1alpha1.RoutingConfig{
-				ProxyConfig: &controllerv1alpha1.Proxy{
-					HttpProxy:  ptr.To(""),
-					HttpsProxy: ptr.To(""),
-					NoProxy:    ptr.To(""),
-				},
-			},
-		},
-		{
-			name: "TLS certificate configmap ref set without proxy does not clobber routing",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-			},
-			existedObjects: []client.Object{
-				&corev1.ConfigMap{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      tls.CheMergedCABundleCertsCMName,
-						Namespace: "eclipse-che",
-					},
-					Data: map[string]string{
-						"ca-bundle.crt": "certificate-data",
-					},
-				},
-			},
-			expectedRoutingConfig: &controllerv1alpha1.RoutingConfig{
-				TLSCertificateConfigmapRef: &controllerv1alpha1.ConfigmapReference{
-					Name:      tls.CheMergedCABundleCertsCMName,
-					Namespace: "eclipse-che",
-				},
-			},
-		},
-		{
-			name: "Both proxy and TLS compose with existing routing config",
-			cheCluster: &chev2.CheCluster{
-				ObjectMeta: metav1.ObjectMeta{
-					Namespace: "eclipse-che",
-					Name:      "eclipse-che",
-				},
-			},
-			existedObjects: []client.Object{
-				&corev1.ConfigMap{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      tls.CheMergedCABundleCertsCMName,
-						Namespace: "eclipse-che",
-					},
-					Data: map[string]string{
-						"ca-bundle.crt": "certificate-data",
-					},
-				},
-				&controllerv1alpha1.DevWorkspaceOperatorConfig{
-					ObjectMeta: metav1.ObjectMeta{
-						Name:      devWorkspaceConfigName,
-						Namespace: "eclipse-che",
-					},
-					TypeMeta: metav1.TypeMeta{
-						Kind:       "DevWorkspaceOperatorConfig",
-						APIVersion: controllerv1alpha1.GroupVersion.String(),
-					},
-					Config: &controllerv1alpha1.OperatorConfiguration{
-						Routing: &controllerv1alpha1.RoutingConfig{
-							DefaultRoutingClass: "che",
-						},
-					},
-				},
-			},
-			httpProxy:  "http://proxy.example.com:3128",
-			httpsProxy: "https://proxy.example.com:3128",
-			expectedRoutingConfig: &controllerv1alpha1.RoutingConfig{
-				DefaultRoutingClass: "che",
-				TLSCertificateConfigmapRef: &controllerv1alpha1.ConfigmapReference{
-					Name:      tls.CheMergedCABundleCertsCMName,
-					Namespace: "eclipse-che",
-				},
-				ProxyConfig: &controllerv1alpha1.Proxy{
-					HttpProxy:  ptr.To(""),
-					HttpsProxy: ptr.To(""),
-					NoProxy:    ptr.To(""),
-				},
-			},
-		},
-	}
-
-	for _, testCase := range testCases {
-		t.Run(testCase.name, func(t *testing.T) {
-			deployContext := test.NewCtxBuilder().WithCheCluster(testCase.cheCluster).WithObjects(testCase.existedObjects...).Build()
-			deployContext.Proxy.HttpProxy = testCase.httpProxy
-			deployContext.Proxy.HttpsProxy = testCase.httpsProxy
-
-			devWorkspaceConfigReconciler := NewDevWorkspaceConfigReconciler()
-			test.EnsureReconcile(t, deployContext, devWorkspaceConfigReconciler.Reconcile)
-
-			dwoc := &controllerv1alpha1.DevWorkspaceOperatorConfig{}
-			err := deployContext.ClusterAPI.Client.Get(
-				context.TODO(),
-				types.NamespacedName{
-					Name:      devWorkspaceConfigName,
-					Namespace: testCase.cheCluster.Namespace,
-				},
-				dwoc,
-			)
-
-			assert.NoError(t, err)
-			assert.Equal(t, testCase.expectedRoutingConfig, dwoc.Config.Routing)
-		})
-	}
-}
